Move .printf to proxy
This means that .printf is also available on fakes, which makes fakes
more useful when used with referee.

See https://github.com/sinonjs/referee-sinon/pull/64
