+++ lib/sinon/proxy.js
@@ -6,6 +6,7 @@ var functionToString = require("./util/core/function-to-string");
 var proxyCall = require("./proxy-call");
 var proxyCallUtil = require("./proxy-call-util");
 var proxyInvoke = require("./proxy-invoke");
+var sinonFormat = require("./util/core/format");
 
 var push = arrayProto.push;
 var forEach = arrayProto.forEach;
@@ -100,6 +101,25 @@ var proxyApi = {
         }
 
         return this.callIds[this.callCount - 1] === proxy.callIds[proxy.callCount - 1] + 1;
+    },
+
+    formatters: require("./spy-formatters"),
+    printf: function(format) {
+        var spyInstance = this;
+        var args = slice(arguments, 1);
+        var formatter;
+
+        return (format || "").replace(/%(.)/g, function(match, specifyer) {
+            formatter = proxyApi.formatters[specifyer];
+
+            if (typeof formatter === "function") {
+                return String(formatter(spyInstance, args));
+            } else if (!isNaN(parseInt(specifyer, 10))) {
+                return sinonFormat(args[specifyer - 1]);
+            }
+
+            return "%" + specifyer;
+        });
     },
 
     resetHistory: function() {
