+++ lib/sinon/spy.js
@@ -10,7 +10,6 @@ var isEsModule = require("./util/core/is-es-module");
 var proxyCallUtil = require("./proxy-call-util");
 var walkObject = require("./util/core/walk-object");
 var wrapMethod = require("./util/core/wrap-method");
-var sinonFormat = require("./util/core/format");
 var valueToString = require("@sinonjs/commons").valueToString;
 
 /* cache references to library methods so that they also can be stubbed without problems */
@@ -32,8 +31,6 @@ function matches(fake, args, strict) {
 
 // Public API
 var spyApi = {
-    formatters: require("./spy-formatters"),
-
     withArgs: function() {
         var args = slice(arguments);
         var matching = pop(this.matchingFakes(args, true));
@@ -73,24 +70,6 @@ var spyApi = {
     matchingFakes: function(args, strict) {
         return filter.call(this.fakes, function(fake) {
             return matches(fake, args, strict);
-        });
-    },
-
-    printf: function(format) {
-        var spyInstance = this;
-        var args = slice(arguments, 1);
-        var formatter;
-
-        return (format || "").replace(/%(.)/g, function(match, specifyer) {
-            formatter = spyApi.formatters[specifyer];
-
-            if (typeof formatter === "function") {
-                return String(formatter(spyInstance, args));
-            } else if (!isNaN(parseInt(specifyer, 10))) {
-                return sinonFormat(args[specifyer - 1]);
-            }
-
-            return "%" + specifyer;
         });
     }
 };
